requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
google-generativeai==0.3.2
Pillow==10.1.0
python-dotenv==1.0.0
gunicorn==21.2.0
markupsafe==2.1.5
itsdangerous==2.1.2
click==8.1.7
blinker==1.7.0
Jinja2==3.1.3
waitress==3.0.0
